feat(enquete): extract menu builder service #1832
